Refactor: basetest (#237)

+++ src/test/java/xyz/npgw/test/common/base/BaseTest.java
@@ -1,14 +1,12 @@
 package xyz.npgw.test.common.base;
 
-import com.google.gson.Gson;
 import com.microsoft.playwright.APIRequestContext;
-import com.microsoft.playwright.APIResponse;
 import com.microsoft.playwright.Browser;
 import com.microsoft.playwright.BrowserContext;
 import com.microsoft.playwright.Page;
 import com.microsoft.playwright.Playwright;
-import com.microsoft.playwright.options.RequestOptions;
 import io.qameta.allure.Allure;
+import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.extern.log4j.Log4j2;
 import org.testng.ITestResult;
@@ -16,51 +14,37 @@ import org.testng.annotations.AfterClass;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Optional;
-import org.testng.annotations.Parameters;
+import org.testng.annotations.Test;
+import xyz.npgw.test.common.ApiContextUtils;
 import xyz.npgw.test.common.BrowserFactory;
 import xyz.npgw.test.common.PlaywrightOptions;
 import xyz.npgw.test.common.ProjectProperties;
-import xyz.npgw.test.common.ProjectUtils;
 import xyz.npgw.test.common.UserRole;
+import xyz.npgw.test.page.AboutBlankPage;
 
 import java.io.IOException;
 import java.lang.reflect.Method;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Map;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 
 @Log4j2
 public abstract class BaseTest {
 
-    private static final String ARTEFACT_DIR = "target/artefact";
-
     private Playwright playwright;
     private Browser browser;
-    private String browserType;
     private BrowserContext context;
+    @Getter(AccessLevel.PROTECTED)
     private Page page;
-    @Getter
-    private APIRequestContext request;
+    @Getter(AccessLevel.PROTECTED)
+    private APIRequestContext apiRequestContext;
 
-    @Parameters("browserType")
     @BeforeClass
-    protected void beforeClass(@Optional("Chromium") String browserType) {
-        this.browserType = browserType.toUpperCase();
-        try {
-            playwright = Playwright.create();
-            browser = BrowserFactory.getBrowser(playwright, this.browserType);
-        } catch (IllegalArgumentException e) {
-            if (playwright != null) {
-                playwright.close();
-            }
-            log.error("Unsupported browser: {}", browserType);
-            System.exit(1);
-        } catch (RuntimeException e) {
-            log.error("Playwright.create() failed: {}", e.getMessage());
-            System.exit(2);
-        }
+    protected void beforeClass() {
+        playwright = Playwright.create();
+        browser = BrowserFactory.getBrowser(playwright);
     }
 
     @BeforeMethod
@@ -71,18 +55,8 @@ public abstract class BaseTest {
             context.tracing().start(PlaywrightOptions.tracingStartOptions());
         }
 
-        APIResponse tokenResponse = playwright.request().newContext().post(
-                ProjectProperties.getBaseUrl() + "/portal-v1/user/token",
-                RequestOptions.create().setData(
-                        Map.of("email", ProjectProperties.getUserEmail(),
-                                "password", ProjectProperties.getUserPassword())));
-        if (tokenResponse.ok()) {
-            log.debug(tokenResponse.text());
-            String idToken = new Gson().fromJson(tokenResponse.text(), TokenResponse.class).token().idToken;
-            request = playwright.request().newContext(PlaywrightOptions.apiContextOptions(idToken));
-        } else {
-            log.error("Retrieve API idToken failed: {}", tokenResponse.statusText());
-            System.exit(5);
+        if (apiRequestContext == null) {
+            apiRequestContext = ApiContextUtils.getApiRequestContext(playwright);
         }
 
         page = context.newPage();
@@ -92,54 +66,49 @@ public abstract class BaseTest {
             try {
                 userRole = UserRole.valueOf((String) args[0]);
             } catch (IllegalArgumentException e) {
-                log.debug("Unknown UserRole, using 'SUPER' as default. {}", e.getMessage());
+                log.debug("Unknown user role, using SUPER");
             }
         }
-        if (userRole != UserRole.GUEST) {
-            Allure.step("Login to the site as %s".formatted(userRole));
-            ProjectUtils.loginAsRole(page, userRole);
+        if (userRole == UserRole.GUEST) {
+            new AboutBlankPage(page).navigate("/");
+        } else {
+            new AboutBlankPage(page).navigate("/").loginAs(userRole);
         }
     }
 
     @AfterMethod(alwaysRun = true)
     protected void afterMethod(Method method, ITestResult testResult) {
-        String testName = ProjectUtils.getTestClassCompleteMethodName(method, testResult);
-        Path traceFilePath = Paths.get(ARTEFACT_DIR, browserType, testName + ".zip");
-        Path videoFilePath = Paths.get(ARTEFACT_DIR, browserType, testName + ".webm");
+        String testName = getTestName(method, testResult);
 
+        long testDuration = (testResult.getEndMillis() - testResult.getStartMillis()) / 1000;
+        log.info("{} {} in {} s", testName, testResult.isSuccess(), testDuration);
+
+        Path videoFilePath = Paths
+                .get(ProjectProperties.getArtefactDir(), ProjectProperties.getBrowserType(), testName + ".webm");
         if (page != null) {
             page.close();
 
-            if (ProjectProperties.isVideoMode()) {
-                page.video().saveAs(videoFilePath);
+            if (ProjectProperties.isVideoMode() && page.video() != null) {
+                if (!testResult.isSuccess()) {
+                    page.video().saveAs(videoFilePath);
+                    addVideoToAllure(testResult, videoFilePath);
+                }
                 page.video().delete();
             }
         }
 
+        Path traceFilePath = Paths
+                .get(ProjectProperties.getArtefactDir(), ProjectProperties.getBrowserType(), testName + ".zip");
         if (context != null) {
             if (ProjectProperties.isTracingMode()) {
-                context.tracing().stop(PlaywrightOptions.tracingStopOptions(traceFilePath));
+                if (testResult.isSuccess()) {
+                    context.tracing().stop();
+                } else {
+                    context.tracing().stop(PlaywrightOptions.tracingStopOptions(traceFilePath));
+                    addTracesToAllure(testResult, traceFilePath);
+                }
             }
             context.close();
-        }
-
-        if (!testResult.isSuccess()) { // && ProjectProperties.isServerRun()) {
-            if (ProjectProperties.isVideoMode()) {
-                try {
-                    Allure.getLifecycle().addAttachment(
-                            "video", "video/webm", "webm", Files.readAllBytes(videoFilePath));
-                } catch (IOException e) {
-                    log.error("Add video to allure failed: {}", e.getMessage());
-                }
-            }
-            if (ProjectProperties.isTracingMode()) {
-                try {
-                    Allure.getLifecycle().addAttachment(
-                            "tracing", "archive/zip", "zip", Files.readAllBytes(traceFilePath));
-                } catch (IOException e) {
-                    log.error("Add traces to allure failed: {}", e.getMessage());
-                }
-            }
         }
     }
 
@@ -148,18 +117,37 @@ public abstract class BaseTest {
         if (browser != null) {
             browser.close();
         }
+        if (apiRequestContext != null) {
+            apiRequestContext.dispose();
+        }
         if (playwright != null) {
             playwright.close();
         }
     }
 
-    protected Page getPage() {
-        return page;
+    private String getTestName(Method method, ITestResult testResult) {
+        String testName = method.getDeclaringClass().getSimpleName() + "/" + method.getName();
+        if (!method.getAnnotation(Test.class).dataProvider().isEmpty()) {
+            testName = "%s(%d)".formatted(testName, testResult.getMethod().getCurrentInvocationCount() - 1);
+        }
+        return testName + new SimpleDateFormat("_MMdd_HHmmss").format(new Date());
     }
 
-    private record Token(String accessToken, int expiresIn, String idToken, String refreshToken, String tokenType) {
+    private void addVideoToAllure(ITestResult testResult, Path path) {
+        try {
+            Allure.getLifecycle()
+                    .addAttachment("video", "video/webm", "webm", Files.readAllBytes(path));
+        } catch (IOException e) {
+            log.error("Add video to allure failed: {}", e.getMessage());
+        }
     }
 
-    private record TokenResponse(String userChallengeType, Token token) {
+    private void addTracesToAllure(ITestResult testResult, Path path) {
+        try {
+            Allure.getLifecycle()
+                    .addAttachment("tracing", "archive/zip", "zip", Files.readAllBytes(path));
+        } catch (IOException e) {
+            log.error("Add traces to allure failed: {}", e.getMessage());
+        }
     }
 }
